word-pattern-replacement/requirements.txt
python-docx
pandas
python-dotenv
pyinstaller
openpyxl
